Relocate dash demo modules under src
