--- services/greeting/__init__.py
@@ -1,5 +1,0 @@
-class GreetingService:
-    """Simple greeting service."""
-
-    def greet(self, name: str) -> str:
-        return f"Hello, {name}!"
