+++ utils/ai_column_mapper.py
@@ -6,7 +6,7 @@ from typing import Dict, List, Optional
 
 import pandas as pd
 
-from components.plugin_adapter import ComponentPluginAdapter
+from yosai_intel_dashboard.src.components.plugin_adapter import ComponentPluginAdapter
 from mapping.models import ColumnRules, load_rules
 
 # ---------------------------------------------------------------------------
