+++ yosai_intel_dashboard/src/app.py
@@ -2,9 +2,9 @@
 
 from dash import Dash, html
 
-from pages import greetings, register_callbacks
-from services.greeting import GreetingService
-from simple_di import ServiceContainer
+from yosai_intel_dashboard.src.pages import greetings, register_callbacks
+from yosai_intel_dashboard.src.services.greeting import GreetingService
+from yosai_intel_dashboard.src.simple_di import ServiceContainer
 
 
 def create_app() -> Dash:
