+++ yosai_intel_dashboard/src/core/app_factory/__init__.py
@@ -6,7 +6,7 @@ import dash_bootstrap_components as dbc
 from dash import dcc, html
 
 # Import Path for building robust file paths
-from components.ui.navbar import create_navbar_layout
+from yosai_intel_dashboard.src.components.ui.navbar import create_navbar_layout
 from core.error_handlers import register_error_handlers
