+++ yosai_intel_dashboard/src/pages/device_learning/callbacks.py
