+++ yosai_intel_dashboard/src/pages/greetings/layout.py
