+++ yosai_intel_dashboard/src/services/upload/core_api.py
@@ -1,6 +1,6 @@
 """Core upload services."""
 
-from components.ui_builder import UploadUIBuilder
+from yosai_intel_dashboard.src.components.ui_builder import UploadUIBuilder
 from validation.file_validator import FileValidator
 
 from .file_processor_service import FileProcessor
