+++ yosai_intel_dashboard/src/simple_di.py
